cloudwatch-logs-api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
boto3==1.34.49
aioboto3==12.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.26.0
orjson==3.9.12
python-multipart==0.0.7
fastapi-pagination==0.12.12
email-validator==2.1.0.post1
moto==4.2.5    # For testing AWS services